+++ News.md
@@ -1,3 +1,9 @@
+0.4.42
+namehandling tests
+
+0.4.42
+fix str.extract code for pandas >=0.18
+
 0.4.41
 Centralize name handling into its won fucntion to be accecessed by split fasta and merge OTUs